Move common data setters of entities to AbstractEntityBuilder
