+++ src/main/java/net/dv8tion/jda/internal/entities/EntityBuilder.java
@@ -18,7 +18,6 @@ package net.dv8tion.jda.internal.entities;
 
 import gnu.trove.map.TLongObjectMap;
 import gnu.trove.map.hash.TLongObjectHashMap;
-import net.dv8tion.jda.api.JDA;
 import net.dv8tion.jda.api.OnlineStatus;
 import net.dv8tion.jda.api.Permission;
 import net.dv8tion.jda.api.audit.ActionType;
@@ -35,7 +34,6 @@ import net.dv8tion.jda.api.entities.channel.ChannelType;
 import net.dv8tion.jda.api.entities.channel.attribute.IThreadContainer;
 import net.dv8tion.jda.api.entities.channel.attribute.IWebhookContainer;
 import net.dv8tion.jda.api.entities.channel.concrete.*;
-import net.dv8tion.jda.api.entities.channel.forums.ForumTag;
 import net.dv8tion.jda.api.entities.channel.middleman.AudioChannel;
 import net.dv8tion.jda.api.entities.channel.middleman.GuildChannel;
 import net.dv8tion.jda.api.entities.channel.middleman.GuildMessageChannel;
@@ -64,7 +62,6 @@ import net.dv8tion.jda.internal.JDAImpl;
 import net.dv8tion.jda.internal.entities.channel.concrete.*;
 import net.dv8tion.jda.internal.entities.channel.mixin.attribute.IInteractionPermissionMixin;
 import net.dv8tion.jda.internal.entities.channel.mixin.attribute.IPermissionContainerMixin;
-import net.dv8tion.jda.internal.entities.channel.mixin.attribute.IPostContainerMixin;
 import net.dv8tion.jda.internal.entities.channel.mixin.middleman.AudioChannelMixin;
 import net.dv8tion.jda.internal.entities.emoji.CustomEmojiImpl;
 import net.dv8tion.jda.internal.entities.emoji.RichCustomEmojiImpl;
@@ -81,7 +78,6 @@ import net.dv8tion.jda.internal.utils.UnlockHook;
 import net.dv8tion.jda.internal.utils.cache.ChannelCacheViewImpl;
 import net.dv8tion.jda.internal.utils.cache.MemberCacheViewImpl;
 import net.dv8tion.jda.internal.utils.cache.SnowflakeCacheViewImpl;
-import net.dv8tion.jda.internal.utils.cache.SortedSnowflakeCacheViewImpl;
 import org.apache.commons.collections4.CollectionUtils;
 import org.apache.commons.collections4.map.CaseInsensitiveMap;
 import org.slf4j.Logger;
@@ -96,10 +92,9 @@ import java.util.*;
 import java.util.function.Function;
 import java.util.function.UnaryOperator;
 import java.util.stream.Collectors;
-import java.util.stream.IntStream;
 import java.util.stream.StreamSupport;
 
-public class EntityBuilder
+public class EntityBuilder extends AbstractEntityBuilder
 {
     public static final Logger LOG = JDALogger.getLog(EntityBuilder.class);
     public static final String MISSING_CHANNEL = "MISSING_CHANNEL";
@@ -119,16 +114,9 @@ public class EntityBuilder
         richGameFields = Collections.unmodifiableSet(tmp);
     }
 
-    protected final JDAImpl api;
-
-    public EntityBuilder(JDA api)
+    public EntityBuilder(JDAImpl api)
     {
-        this.api = (JDAImpl) api;
-    }
-
-    public JDAImpl getJDA()
-    {
-        return api;
+        super(api);
     }
 
     public SelfUser createSelfUser(DataObject self)
@@ -1156,11 +1144,7 @@ public class EntityBuilder
     {
         final long id = json.getLong("id");
         final CategoryImpl channel = new CategoryImpl(id, guild);
-
-        channel
-                .setName(json.getString("name"))
-                .setPosition(json.getInt("position"));
-
+        configureCategory(json, channel);
         createChannelInteractionPermissions(channel, json);
         return channel;
     }
@@ -1202,17 +1186,7 @@ public class EntityBuilder
     {
         final long id = json.getLong("id");
         TextChannelImpl channel = new TextChannelImpl(id, guildObj);
-
-        channel
-                .setParentCategory(json.getLong("parent_id", 0))
-                .setLatestMessageIdLong(json.getLong("last_message_id", 0))
-                .setName(json.getString("name"))
-                .setTopic(json.getString("topic", null))
-                .setPosition(json.getInt("position"))
-                .setNSFW(json.getBoolean("nsfw"))
-                .setDefaultThreadSlowmode(json.getInt("default_thread_rate_limit_per_user", 0))
-                .setSlowmode(json.getInt("rate_limit_per_user", 0));
-
+        configureTextChannel(json, channel);
         createChannelInteractionPermissions(channel, json);
         return channel;
     }
@@ -1254,15 +1228,7 @@ public class EntityBuilder
     {
         final long id = json.getLong("id");
         NewsChannelImpl channel = new NewsChannelImpl(id, guildObj);
-
-        channel
-                .setParentCategory(json.getLong("parent_id", 0))
-                .setLatestMessageIdLong(json.getLong("last_message_id", 0))
-                .setName(json.getString("name"))
-                .setTopic(json.getString("topic", null))
-                .setPosition(json.getInt("position"))
-                .setNSFW(json.getBoolean("nsfw"));
-
+        configureNewsChannel(json, channel);
         createChannelInteractionPermissions(channel, json);
         return channel;
     }
@@ -1303,20 +1269,7 @@ public class EntityBuilder
     {
         final long id = json.getLong("id");
         VoiceChannelImpl channel = new VoiceChannelImpl(id, guild);
-
-        channel
-                .setParentCategory(json.getLong("parent_id", 0))
-                .setLatestMessageIdLong(json.getLong("last_message_id", 0))
-                .setName(json.getString("name"))
-                .setStatus(json.getString("status", ""))
-                .setPosition(json.getInt("position"))
-                .setUserLimit(json.getInt("user_limit"))
-                .setNSFW(json.getBoolean("nsfw"))
-                .setBitrate(json.getInt("bitrate"))
-                .setRegion(json.getString("rtc_region", null))
-//            .setDefaultThreadSlowmode(json.getInt("default_thread_rate_limit_per_user", 0))
-                .setSlowmode(json.getInt("rate_limit_per_user", 0));
-
+        configureVoiceChannel(json, channel);
         createChannelInteractionPermissions(channel, json);
         return channel;
     }
@@ -1357,18 +1310,7 @@ public class EntityBuilder
     {
         final long id = json.getLong("id");
         final StageChannelImpl channel = new StageChannelImpl(id, guild);
-        channel
-                .setParentCategory(json.getLong("parent_id", 0))
-                .setLatestMessageIdLong(json.getLong("last_message_id", 0))
-                .setName(json.getString("name"))
-                .setPosition(json.getInt("position"))
-                .setBitrate(json.getInt("bitrate"))
-                .setUserLimit(json.getInt("user_limit", 0))
-                .setNSFW(json.getBoolean("nsfw"))
-                .setRegion(json.getString("rtc_region", null))
-//            .setDefaultThreadSlowmode(json.getInt("default_thread_rate_limit_per_user", 0))
-                .setSlowmode(json.getInt("rate_limit_per_user", 0));
-
+        configureStageChannel(json, channel);
         createChannelInteractionPermissions(channel, json);
         return channel;
     }
@@ -1437,31 +1379,7 @@ public class EntityBuilder
         final long id = json.getUnsignedLong("id");
         final ChannelType type = ChannelType.fromId(json.getInt("type"));
         ThreadChannelImpl channel = new ThreadChannelImpl(id, guild, type);
-
-        DataObject threadMetadata = json.getObject("thread_metadata");
-
-        if (!json.isNull("applied_tags") && api.isCacheFlagSet(CacheFlag.FORUM_TAGS))
-        {
-            DataArray array = json.getArray("applied_tags");
-            channel.setAppliedTags(IntStream.range(0, array.length()).mapToLong(array::getUnsignedLong));
-        }
-
-        channel
-                .setName(json.getString("name"))
-                .setFlags(json.getInt("flags", 0))
-                .setOwnerId(json.getLong("owner_id"))
-                .setMemberCount(json.getInt("member_count"))
-                .setMessageCount(json.getInt("message_count"))
-                .setTotalMessageCount(json.getInt("total_message_count", 0))
-                .setLatestMessageIdLong(json.getLong("last_message_id", 0))
-                .setSlowmode(json.getInt("rate_limit_per_user", 0))
-                .setLocked(threadMetadata.getBoolean("locked"))
-                .setArchived(threadMetadata.getBoolean("archived"))
-                .setInvitable(threadMetadata.getBoolean("invitable"))
-                .setArchiveTimestamp(Helpers.toTimestamp(threadMetadata.getString("archive_timestamp")))
-                .setCreationTimestamp(threadMetadata.isNull("create_timestamp") ? 0 : Helpers.toTimestamp(threadMetadata.getString("create_timestamp")))
-                .setAutoArchiveDuration(ThreadChannel.AutoArchiveDuration.fromKey(threadMetadata.getInt("auto_archive_duration")));
-
+        configureThreadChannel(json, channel);
         createChannelInteractionPermissions(channel, json);
         return channel;
     }
@@ -1519,27 +1437,7 @@ public class EntityBuilder
     {
         final long id = json.getLong("id");
         final ForumChannelImpl channel = new ForumChannelImpl(id, guild);
-
-        if (api.isCacheFlagSet(CacheFlag.FORUM_TAGS))
-        {
-            DataArray tags = json.getArray("available_tags");
-            for (int i = 0; i < tags.length(); i++)
-                createForumTag(channel, tags.getObject(i), i);
-        }
-
-        channel
-                .setParentCategory(json.getLong("parent_id", 0))
-                .setFlags(json.getInt("flags", 0))
-                .setDefaultReaction(json.optObject("default_reaction_emoji").orElse(null))
-                .setDefaultSortOrder(json.getInt("default_sort_order", -1))
-                .setDefaultLayout(json.getInt("default_forum_layout", -1))
-                .setName(json.getString("name"))
-                .setTopic(json.getString("topic", null))
-                .setPosition(json.getInt("position"))
-                .setDefaultThreadSlowmode(json.getInt("default_thread_rate_limit_per_user", 0))
-                .setSlowmode(json.getInt("rate_limit_per_user", 0))
-                .setNSFW(json.getBoolean("nsfw"));
-
+        configureForumChannel(json, channel);
         createChannelInteractionPermissions(channel, json);
         return channel;
     }
@@ -1579,50 +1477,9 @@ public class EntityBuilder
     {
         final long id = json.getLong("id");
         final MediaChannelImpl channel = new MediaChannelImpl(id, guild);
-
-        if (api.isCacheFlagSet(CacheFlag.FORUM_TAGS))
-        {
-            DataArray tags = json.getArray("available_tags");
-            for (int i = 0; i < tags.length(); i++)
-                createForumTag(channel, tags.getObject(i), i);
-        }
-
-        channel
-                .setParentCategory(json.getLong("parent_id", 0))
-                .setFlags(json.getInt("flags", 0))
-                .setDefaultReaction(json.optObject("default_reaction_emoji").orElse(null))
-                .setDefaultSortOrder(json.getInt("default_sort_order", -1))
-                .setName(json.getString("name"))
-                .setTopic(json.getString("topic", null))
-                .setPosition(json.getInt("position"))
-                .setDefaultThreadSlowmode(json.getInt("default_thread_rate_limit_per_user", 0))
-                .setSlowmode(json.getInt("rate_limit_per_user", 0))
-                .setNSFW(json.getBoolean("nsfw"));
-
+        configureMediaChannel(json, channel);
         createChannelInteractionPermissions(channel, json);
         return channel;
-    }
-
-    public ForumTagImpl createForumTag(IPostContainerMixin<?> channel, DataObject json, int index)
-    {
-        final long id = json.getUnsignedLong("id");
-        SortedSnowflakeCacheViewImpl<ForumTag> cache = channel.getAvailableTagCache();
-        ForumTagImpl tag = (ForumTagImpl) cache.get(id);
-
-        if (tag == null)
-        {
-            try (UnlockHook lock = cache.writeLock())
-            {
-                tag = new ForumTagImpl(id);
-                cache.getMap().put(id, tag);
-            }
-        }
-
-        tag.setName(json.getString("name"))
-           .setModerated(json.getBoolean("moderated"))
-           .setEmoji(json)
-           .setPosition(index);
-        return tag;
     }
 
     public PrivateChannel createPrivateChannel(DataObject json)
